requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.4
django-filter==24.2
django-oauth-toolkit==2.3.0
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.15.1
drf-social-oauth2==2.1.3
ecdsa==0.19.0
idna==3.6
jwcrypto==1.5.6
Markdown==3.6
oauthlib==3.2.2
pyasn1==0.6.0
pycparser==2.22
PyJWT==2.8.0
python-jose==3.3.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.11.0
urllib3==2.2.1
